// testdata/compat/ts/person_service.pb.ts
// Code generated by gcode. DO NOT EDIT.
// source: person_service.proto

export interface CreatePersonResponse {
  id: string
}

export interface GetPersonRequest {
  id: string
}

export interface GetPersonResponse {
  name: string
  age: number
}

export interface UpdatePersonResponse {
  ok: boolean
}

export interface DeletePersonRequest {
  id: string
}

export interface DeletePersonResponse {
  ok: boolean
}

export const GetPersonRequestRules = {
  id: { required: false, type: "string", minLength: 1, maxLength: 64 }
} as const

export const DeletePersonRequestRules = {
  id: { required: false, type: "string", minLength: 1, maxLength: 64 }
} as const