// testdata/compat/ts/all_types.create.pb.ts
// Code generated by gcode. DO NOT EDIT.
// source: all_types.create.proto

export interface AllScalarsCreate {
  fSint32?: number
  fSint64?: string
  fSfixed32?: number
  fSfixed64?: string
  fFixed32?: number
  fFixed64?: string
  fUint32?: number
  fUint64?: string
  fFloat?: number
  fBytes?: string
}